# Make 1D integer sorting parallel