--- a/.gitmodules
+++ b/.gitmodules
@@ -105,7 +105,7 @@
 [submodule "third_party/fbgemm"]
     ignore = dirty
     path = third_party/fbgemm
-    url = https://github.com/pytorch/fbgemm
+    url = https://github.com/DamianSzwichtenberg/fbgemm
 [submodule "third_party/foxi"]
     ignore = dirty
     path = third_party/foxi