--- a/aten/src/ATen/native/cpu/SortingKernel.cpp
+++ b/aten/src/ATen/native/cpu/SortingKernel.cpp
@@ -1,13 +1,19 @@
 #define TORCH_ASSERT_NO_OPERATORS
+
+#include <limits>
+
 #include <ATen/native/Sorting.h>
 #include <ATen/core/TensorBase.h>
 #include <ATen/Dispatch.h>
 #include <ATen/Parallel.h>
 #include <ATen/NumericUtils.h>
 #include <ATen/TensorIterator.h>
+#include <ATen/cpu/vec/functional.h>
+#include <ATen/cpu/vec/vec.h>
 #include <ATen/native/StridedRandomAccessor.h>
 #include <ATen/native/CompositeRandomAccessor.h>
 #include <ATen/native/TopKImpl.h>
+#include <ATen/native/cpu/radix_sort.h>
 #include <c10/core/WrapDimMinimal.h>
 #include <c10/util/irange.h>
 
@@ -83,6 +89,41 @@ struct KeyValueCompDesc {
   }
 };
 
+static void parallel_sort1d_kernel(
+    const TensorBase& values,
+    const TensorBase& indices) {
+  // this kernel does not care about `stable` parameter as radix sort
+  // used here is a stable sorting algorithm
+  AT_DISPATCH_INTEGRAL_TYPES(values.scalar_type(), "parallel_sort1d_kernel", [&] {
+    int64_t elements = values.numel();
+    scalar_t* keys = values.data_ptr<scalar_t>();
+    int64_t* vals = indices.data_ptr<int64_t>();
+    std::vector<scalar_t> tmp_keys(elements);
+    std::vector<int64_t> tmp_vals(elements);
+    scalar_t* sorted_keys = nullptr;
+    int64_t* sorted_vals = nullptr;
+    std::tie(sorted_keys, sorted_vals) = radix_sort_parallel(
+        keys,
+        vals,
+        tmp_keys.data(),
+        tmp_vals.data(),
+        elements,
+        std::numeric_limits<scalar_t>::max(),
+        values.scalar_type() != ScalarType::Byte);
+
+    const bool sorted_in_place = keys == sorted_keys;
+    if (!sorted_in_place) {
+      const auto common_size = values.numel();
+      const int num_threads = at::get_num_threads();
+      at::parallel_for(0, common_size, at::internal::GRAIN_SIZE / num_threads, [&](int64_t begin, int64_t end) {
+        const auto job_size = end - begin;
+        vec::map([](vec::Vectorized<scalar_t> x) -> vec::Vectorized<scalar_t> { return x; }, keys + begin, sorted_keys + begin, job_size);
+        vec::map([](vec::Vectorized<int64_t> x) -> vec::Vectorized<int64_t> { return x; }, vals + begin, sorted_vals + begin, job_size);
+      });
+    }
+  });
+}
+
 static void sort_kernel(
     const TensorBase& self,
     const TensorBase& values,
@@ -95,6 +136,14 @@ static void sort_kernel(
   if (self.stride(dim) == 0) {
     // check if stride is zero
     // https://github.com/pytorch/pytorch/issues/91420
+    return;
+  }
+  // TODO(dszwicht): Should we add here check for `stable` param?
+  // Radix sort is a stable sorting algorithm.
+  if (values.dim() == 1 && values.numel() >= at::internal::GRAIN_SIZE &&
+      at::isIntegralType(values.scalar_type(), /*includeBool=*/false) &&
+      is_radix_sort_available() && !descending) {
+    parallel_sort1d_kernel(values, indices);
     return;
   }
   _dim_apply(