In GNN workloads we use `torch.argsort` to calculate permutation from CSR to CSC sparse matrix storage format. Till now, sorting one-dimensional data was performed sequentially. This change reuses radix sort from [fbgemm](https://github.com/dszwicht/FBGEMM/pull/1) and makes `torch.(arg)sort` work in parallel.

Performance measurements (measured on ICX platform - 40C):
```
| Size        | Before [s] | Now [s] | Speedup |
|-------------|------------|---------|---------|
| 10^5        | 0.0107     | 0.002   | 5.35x   |
| 4 * 10^5    | 0.0402     | 0.0059  | 6.81x   |
| 16 * 10^5   | 0.1756     | 0.0159  | 11.04x  |
| 64 * 10^5   | 0.7659     | 0.0626  | 12.23x  |
| 256 * 10^5  | 3.2334     | 0.2636  | 12.26x  |
| 1024 * 10^5 | 13.356     | 1.0283  | 12.98x  |
```